+++ gipit-back/app/api/management/route.ts
@@ -1,9 +1,10 @@
 import { PrismaClient } from '@prisma/client';
 import { NextRequest, NextResponse } from 'next/server';
+import { verifyToken } from '../middleware'; 
 
 const prisma = new PrismaClient();
 
-// Obtener todas las jefaturas (GET)
+// Obtener todos los registros de management (GET)
 /**
  * @swagger
  * /management:
@@ -19,8 +20,14 @@ const prisma = new PrismaClient();
  *               items:
  *                 $ref: '#/components/schemas/Management'
  */
-export async function GET() {
+export async function GET(req: NextRequest) {
   try {
+    // Verify JWT token
+    const verificationResult = verifyToken(req);
+    if (!verificationResult.valid) {
+      return NextResponse.json({ error: verificationResult.error }, { status: 403 });
+    }
+
     const managements = await prisma.management.findMany();
     return NextResponse.json(managements);
   } catch (error) {
@@ -28,7 +35,7 @@ export async function GET() {
   }
 }
 
-// Crear una nueva jefatura (POST)
+// Crear un nuevo registro de management (POST)
 /**
  * @swagger
  * /management:
@@ -48,7 +55,14 @@ export async function GET() {
  */
 export async function POST(req: NextRequest) {
   const { company_id, name, description } = await req.json();
+
   try {
+
+    const verificationResult = verifyToken(req);
+    if (!verificationResult.valid) {
+      return NextResponse.json({ error: verificationResult.error }, { status: 403 });
+    }
+
     const management = await prisma.management.create({
       data: { company_id, name, description },
     });
@@ -58,24 +72,38 @@ export async function POST(req: NextRequest) {
   }
 }
 
-// Actualizar una jefatura (PUT)
+// Actualizar un registro de management (PUT)
 export async function PUT(req: NextRequest) {
   const { id, name, description } = await req.json();
+
   try {
+
+    const verificationResult = verifyToken(req);
+    if (!verificationResult.valid) {
+      return NextResponse.json({ error: verificationResult.error }, { status: 403 });
+    }
+
     const management = await prisma.management.update({
       where: { id: parseInt(id) },
       data: { name, description },
     });
     return NextResponse.json(management);
   } catch (error) {
-    return NextResponse.json({ error: `Error updating management -${error}` }, { status: 500 });
+    return NextResponse.json({ error: `Error updating management - ${error}` }, { status: 500 });
   }
 }
 
-// Eliminar una jefatura (DELETE)
+// Eliminar un registro de management (DELETE)
 export async function DELETE(req: NextRequest) {
   const { id } = await req.json();
+
   try {
+
+    const verificationResult = verifyToken(req);
+    if (!verificationResult.valid) {
+      return NextResponse.json({ error: verificationResult.error }, { status: 403 });
+    }
+
     await prisma.management.delete({ where: { id: parseInt(id) } });
     return NextResponse.json({ message: 'Management deleted' }, { status: 204 });
   } catch (error) {
