+++ gipit-back/app/api/post_sales_activities/[id]/route.ts
@@ -1,5 +1,6 @@
 import { PrismaClient } from '@prisma/client';
 import { NextRequest, NextResponse } from 'next/server';
+import { verifyToken } from '../../middleware'; 
 
 const prisma = new PrismaClient();
 
@@ -29,6 +30,12 @@ export async function GET(req: NextRequest, { params }: { params: { id: string }
   const { id } = params;
 
   try {
+
+    const verificationResult = verifyToken(req);
+    if (!verificationResult.valid) {
+      return NextResponse.json({ error: verificationResult.error }, { status: 403 });
+    }
+
     const postSalesActivity = await prisma.post_sales_activities.findUnique({
       where: { id: parseInt(id) },
     });
@@ -42,7 +49,6 @@ export async function GET(req: NextRequest, { params }: { params: { id: string }
     return NextResponse.json({ error: `Error - ${error}` }, { status: 500 });
   }
 }
-
 
 /**
  * @swagger
@@ -72,6 +78,12 @@ export async function PUT(req: NextRequest, { params }: { params: { id: string }
   const { candidate_management_id, benefit, description, date, associated_cost } = await req.json();
 
   try {
+
+    const verificationResult = verifyToken(req);
+    if (!verificationResult.valid) {
+      return NextResponse.json({ error: verificationResult.error }, { status: 403 });
+    }
+
     const updatedPostSalesActivity = await prisma.post_sales_activities.update({
       where: { id: parseInt(id) },
       data: {
@@ -89,7 +101,6 @@ export async function PUT(req: NextRequest, { params }: { params: { id: string }
     return NextResponse.json({ error: `Error - ${error}` }, { status: 500 });
   }
 }
-
 
 /**
  * @swagger
@@ -112,13 +123,18 @@ export async function DELETE(req: NextRequest, { params }: { params: { id: strin
   const { id } = params;
 
   try {
+
+    const verificationResult = verifyToken(req);
+    if (!verificationResult.valid) {
+      return NextResponse.json({ error: verificationResult.error }, { status: 403 });
+    }
+
     await prisma.post_sales_activities.delete({
       where: { id: parseInt(id) },
     });
 
     return NextResponse.json({ message: 'Actividad de ventas eliminada con éxito' });
   } catch (error) {
-    
     return NextResponse.json({ error: `Error - ${error}` }, { status: 500 });
   }
 }
